Changes, bug fixes and improvements.
- i'd like to think at least.

so many memory leaks and bugs out there to be squashed, one day i'll
figure the right type of memory management scheme out and implement it.

- implemented ObjectArray

- changed and removed all references to ReturnResult and swapped them
for Object* references. Code changed to make that happen. It isn't ideal
as I just got it working from the old code base. Needs refining soon,
because of redundant use of Object and Object*'s among other things.

- because of this, many results and underlying types have been moved
around and worked wtih.

AUTO - GENERATED files have been hand coded, so the code generators need
updating before we regerenate code and replace hand written things.

Other stuff i'm too tired to mention.

+++ src/TokenArray.c
@@ -2,12 +2,15 @@
 #include <stdio.h>
 #include "Token.h"
 #include "TokenArray.h"
+
+extern TokenArray* releaseTokenArrayReference(TokenArray*);
 #define INIT_SIZE 5
 
 void init_TokenArray(TokenArray* array){
         array->Tokens = NULL;
         array->size = 0;
         array->used = 0;
+        array->owner_references =1;
         array->addElementToArray = &addElementToTokenArray;
         array->deleteArray = &delete_TokenArray;
         array->getElementInArrayAt =&getTokeninArrayAt;
@@ -35,8 +38,13 @@ void addElementToTokenArray(TokenArray* array,Token* element){
 void delete_TokenArray(TokenArray* array){
     if(array){
         int i;
-        for(i =0; i <array->used;i++){
-            delete_Token(&array->Tokens[i]);
+        if(array->owner_references<=1){
+        	for(i =0; i <array->used;i++){
+        		delete_Token(&array->Tokens[i]);
+        	}
+        }
+        else{
+        	releaseTokenArrayReference(array);
         }
     }
 }
